% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmap.R
\name{coverage_heatmap}
\alias{coverage_heatmap}
\title{Coverage heatmap}
\usage{
coverage_heatmap(data, popmap = NULL, title = NULL,
  males.color = "dodgerblue3", females.color = "red3",
  coverage.palette = c("white", "royalblue2", "black", "gold2", "red3"))
}
\arguments{
\item{data}{A table of coverage obtained with the \code{\link{load_coverage_table}} function.}

\item{popmap}{A population map obtained with the \code{\link{load_population_map}} function.}

\item{title}{Plot title.}

\item{males.color}{If a popmap is specified, sets the color of male individual names on the horizontal axis.}

\item{females.color}{If a popmap is specified, sets the color of female individual names on the horizontal axis.}

\item{coverage.palette}{Color palette for coverage. The value should be a vector of length 5 corresponding to the following intervals/values:
(0, 1 : mean, mean : 3rd quartile, 3rd quartile : (max - 1), max)}
}
\value{
A coverage heatmap stored in a ggplot object.
}
\description{
Generates a heatmap of coverage for a subset of sequences. In the resulting heatmap, the color of a tile at coordinates (x, y)
indicates the coverage of the sequence y in the individual x.
}
\examples{
heatmap = coverage_heatmap(data, popmap=popmap,
                           title="Individuals and sequences clustering based on coverage",
                           males.color="blue", females.color="red",
                           coverage.palette=c("white", "green", "black", "red", "red3"))
}
