% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_distribution_heatmap.R
\name{generate_color_palette}
\alias{generate_color_palette}
\title{Generate color palette}
\usage{
generate_color_palette(color.scale.bins, color.scale.colors)
}
\arguments{
\item{color.scale.bins}{A vector of values to use as bins in the color palette.}

\item{color.scale.colors}{A vector of two colors used to create the color palette gradient.}
}
\value{
A color palette vector.
}
\description{
Generate a palette vector from the specified bins and colors.
}
