% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contig_lengths.R
\name{load_contig_lengths}
\alias{load_contig_lengths}
\title{Load a contig lengths file}
\usage{
load_contig_lengths(input_file_path, group.unplaced = TRUE)
}
\arguments{
\item{input_file_path}{Path to a contig lengths file.}
}
\value{
A vector with contigs as names and lengths as values.
}
\description{
Loads a table of contig lengths.
}
\examples{
lengths = load_contig_lengths(input_file_path)
}
