% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_coverage_table.R
\name{load_coverage_table}
\alias{load_coverage_table}
\title{Coverage table loader}
\usage{
load_coverage_table(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a table of coverage.}
}
\value{
A data frame containing the coverage data.
}
\description{
Load a table of coverage obtained with RADSex subset, RADSex significant_sequences, or RADSex group_loci.
}
\examples{
data = load_coverage_table(input_file_path)
}
