% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contig_names_file.R
\name{load_names_file}
\alias{load_names_file}
\title{Load a contig names file}
\usage{
load_names_file(input_file_path, group.unplaced = TRUE)
}
\arguments{
\item{input_file_path}{Path to a contig names file}
}
\value{
A vector with contigs as names and contig names as values.
}
\description{
Loads a table of contig names.
}
\examples{
lengths = load_names_file(input_file_path)
}
