% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_popmap.R
\name{load_popmap}
\alias{load_popmap}
\title{Popmap loader}
\usage{
load_popmap(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a population map.}
}
\value{
A vector with individuals as names and sex as values.
}
\description{
Load a population map.
}
\examples{
popmap = load_popmap(input_file_path)
}
