% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sex_distribution_table.R
\name{load_sex_distribution_table}
\alias{load_sex_distribution_table}
\title{Load a sex distribution table}
\usage{
load_sex_distribution_table(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a table of distribution of sequences between sexes.}
}
\value{
A list with the following elements:
\item{data}{A data frame of the distribution of sequences between sexes.}
\item{n_males}{Number of males in the population}
\item{n_females}{Number of females in the population}
}
\description{
Loads a table of distribution of sequences between sexes obtained with RADSex sex_distribution.
}
\examples{
data <- load_sex_distribution_table("sex_distribution_table.tsv")
}
