% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_circular_plot.R
\name{mapping_circular_plot}
\alias{mapping_circular_plot}
\title{Mapping circular plot}
\usage{
mapping_circular_plot(data, highlight = NULL, zoom.highlights = FALSE,
  zoom.ratio = 2, zoom.suffix = " (zoom)", base.color = "white",
  highlight.color = "grey80", point.size = 0.5,
  color.sex.bias = TRUE, sex.bias.palette = c("firebrick1", "black",
  "dodgerblue2"), color.unmapped = TRUE, unmapped.palette = c(`0` =
  "dodgerblue3", `1` = "goldenrod1", `2` = "grey30"),
  signif.threshold = 0.05, sector.titles.expand = 1.9)
}
\arguments{
\item{data}{A list of mapping results obtained with the \code{\link{load_mapping_results}} function.}

\item{zoom.highlights}{If TRUE, highlighted scaffolds will be zoomed on at the top of the plot (default FALSE).}

\item{zoom.ratio}{Zooming factor for highlighted scaffolds, i.e. a size multiplier (default 2)}

\item{zoom.suffix}{A suffix to add after the name of a zoomed scaffold (default " (zoom)")}

\item{base.color}{Background color of a standard sector of the plot (default "white").}

\item{highlight.color}{Background color of a highlighted sector of the plot (default "grey80").}

\item{point.size}{Size of a point in the plot (default 0.5)}

\item{color.sex.bias}{If TRUE, points on the sex-bias track will be colored according to sex.bias.palette (default TRUE).}

\item{sex.bias.palette}{A vector of three colors defining the sex-bias track palette: female-biased, neutral, male-biased. (default c("firebrick1", "black", "dodgerblue2"))}

\item{color.unmapped}{If TRUE, unmapped scaffolds will be colored with alternating colors, similar to a manhattan plot (default TRUE).}

\item{unmapped.palette}{A named vector of three colors: "0" = alternating color 1, "1" = alternating color2, and "2" = color for mapped scaffolds (default c("0"="dodgerblue3", "1"="goldenrod1", "2"="grey30")).}

\item{signif.threshold}{Significance threshold for association with sex (default 0.05).}

\item{highlights}{A vector of scaffolds to highlight: c("LG1", "LG7", "scaffold_8") ... (default NULL).}

\item{sector.title.expand}{A factor that controls the distance between sector titles and sector top axes (default 2).}
}
\description{
Generates a circular plot of RADSex mapping results in which each sector represents a linkage group and
the x-axis represents the position on the linkage group. The y-axis on the first track shows the sex-bias of a sequence, and the second track shows the
probability of association with sex of a sequence.
}
\examples{

data <- load_mapping_results("mapping_results.tsv", "contig_lengths.tsv", contig_names = "contig_names.tsv", plot.unplaced = FALSE)

mapping_circular_plot(data,
                      highlight = NULL, zoom.highlights = FALSE, zoom.ratio = 2, zoom.suffix = " (zoom)",
                      base.color = "white", highlight.color = "grey80", point.size = 0.5,
                      color.sex.bias = TRUE, sex.bias.palette = c("firebrick1", "black", "dodgerblue2"),
                      color.unmapped = TRUE, unmapped.palette = c("0"="dodgerblue3", "1"="goldenrod1", "2"="grey30"),
                      signif.threshold = 0.05, sector.titles.expand = 2)
}
