% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_contig.R
\name{mapping_contig}
\alias{mapping_contig}
\title{Mapping contig}
\usage{
mapping_contig(data, contig, region = NULL, title = NULL,
  signif.threshold = 0.05, point.size = 1.5, color.sex.bias = TRUE,
  sex.bias.palette = c("firebrick1", "grey10", "dodgerblue2"),
  association.color = "grey20", significance.line.color = "black",
  significance.line.type = 2, significance.text.position = c(0.05,
  0.05))
}
\arguments{
\item{data}{A list of mapping results obtained with the \code{\link{load_mapping_results}} function.}

\item{contig}{Name of the contig to plot.}

\item{region}{A vector of two integers specifying the region of the contig to plot (default NULL).}

\item{title}{Title of the plot (default NULL).}

\item{signif.threshold}{Significance threshold for association with sex (default 0.05).}

\item{point.size}{Size of a point in the plot (default 1.5).}

\item{color.sex.bias}{If TRUE, points on the sex-bias track will be colored according to sex.bias.palette (default TRUE).}

\item{sex.bias.palette}{A vector of three colors defining the sex-bias track palette: female-biased, neutral, male-biased. (default c("firebrick1", "black", "dodgerblue2"))}

\item{association.color}{Color of the points in association with sex plot (default "grey20").}

\item{significance.line.color}{Color for significance line, set to NULL for no line (default "black").}

\item{significance.line.type}{Linetype for the significance line, as usually defined in R (default 2).}

\item{significance.text.position}{X and Y axis offset for the significance text, as fractions of total axis length (default c(0.05, 0.05)).}
}
\description{
Generates a linear plot of RADSex mapping results for a specified contig. The resulting figure contains two plot: the top plot shows sex-bias against
position on the contig, and the bottom plot shows probability of association with sex against position on the contig.
A specific region of the contig can also be plotted.
}
\examples{
data <- load_mapping_results("mapping_results.tsv", contig_lengths_file_path = 'contig_lengths.tsv',
                             contig_names_file_path = 'contig_names.tsv')
#' contig_plot <- mapping_contig(data, "LG05", region = c(1500000, 2500000), point.size = 1, color.sex.bias = FALSE)
}
