% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{mapping_manhattan_plot}
\alias{mapping_manhattan_plot}
\title{Mapping manhattan plot}
\usage{
mapping_manhattan_plot(data, point.size = 0.5, signif.threshold = 0.05,
  point.palette = c("dodgerblue3", "darkgoldenrod2"),
  background.palette = c("grey85", "grey100"),
  significance.line.color = "black", significance.line.type = 2,
  significance.text.position = c(0.05, 0.05))
}
\arguments{
\item{data}{A list of mapping results obtained with the \code{\link{load_mapping_results}} function.}

\item{point.size}{Size of a point in the plot (default 0.5)}

\item{signif.threshold}{Significance threshold for association with sex (default 0.05).}

\item{point.palette}{Color palette for the dots (default c("dodgerblue3", "darkgoldenrod2"))}

\item{background.palette}{Color palette for the background (default c("grey85", "grey100"))}

\item{significance.line.color}{Color for significance line, set to NULL for no line (default "black").}

\item{significance.line.type}{Linetype for the significance line, as usually defined in R (default 2).}

\item{significance.text.position}{X and Y axis offset for the significance text, as fractions of total axis length (default c(0.05, 0.05)).}
}
\description{
Generates a manhattan plot of association with sex from RADSex mapping results.
}
\examples{

data <- load_mapping_results("mapping_results.tsv", "contig_lengths.tsv", contig_names = "contig_names.tsv", plot.unplaced = FALSE)

 mapping_manhattan_plot(data, point.size = 0.5, signif.threshold = 0.05,
                        point.palette = c("dodgerblue3", "darkgoldenrod2"),
                        background.palette = c("grey85", "grey100"),
                        significance.line.color = "black",
                        significance.line.type = 2,
                        significance.text.position = c(0.05, 0.05)
}
