% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contig.R
\name{plot_contig}
\alias{plot_contig}
\title{Mapping contig}
\usage{
plot_contig(mapping_file_path, contig_lengths_file_path, contig,
  region = NULL, chromosomes_names_file_path = NULL, title = NULL,
  output_file_path = NULL, width = 12, height = 8, dpi = 300,
  color.sex.bias = TRUE, sex.bias.palette = c("firebrick1", "grey10",
  "dodgerblue2"), signif.threshold = 0.05, point.size = 1.5)
}
\arguments{
\item{mapping_file_path}{Path to a mapping results file generated by RADSex map.}

\item{contig_lengths_file_path}{Path to a contig lengths file generated by RADSex map.}

\item{contig}{Name of the contig to plot.}

\item{region}{A vector of two integers specifying the region of the contig to plot (default NULL).}

\item{chromosomes_names_file_path}{Path to a chromosomes names file, i.e. a tabulated file with name in the reference genome file as the first column and
corresponding chromosome name as the second column. If the chromosomes names in the reference genome file start with "LG", "NC", or "CHR" (case unsensitive),
chromosomes can be detected automatically (default NULL).}

\item{title}{Title of the plot (default NULL).}

\item{output_file_path}{Path to the plot output file. If the output file is not specified, the circular plot will be plotted in the default device (default NULL).}

\item{width}{Width of the output file in pixels (default 1400).}

\item{height}{Height of the output file in pixels (default 1400).}

\item{color.sex.bias}{If TRUE, points on the sex-bias track will be colored according to sex.bias.palette (default TRUE).}

\item{sex.bias.palette}{A vector of three colors defining the sex-bias track palette: female-biased, neutral, male-biased. (default c("firebrick1", "black", "dodgerblue2"))}

\item{signif.threshold}{Significance threshold for association with sex (default 0.05).}

\item{point.size}{Size of a point in the plot (default 1.5).}

\item{res}{Resolution of the output file in \% (default 100).}
}
\description{
Generates a linear plot of RADSex mapping results for a specified contig. The resulting figure contains two plot: the top plot shows sex-bias against
position on the contig, and the bottom plot shows probability of association with sex against position on the contig.
A specific region of the contig can also be plotted.
}
\examples{

plot_contig("mapping_results.tsv", "contig_lengths.tsv", "LG01", region = c(10000000, 15000000),
            output_file_path = "mapping_results.png", width = 10, height = 7)
}
