% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{plot_genome}
\alias{plot_genome}
\title{Mapping circular plot}
\usage{
plot_genome(mapping_file_path, contig_lengths_file_path,
  chromosomes_names_file_path = chromosomes_names_file_path,
  plot.unplaced = TRUE, output_file_path = NULL, width = 1400,
  height = 1400, res = 100, highlight = NULL, zoom.highlights = FALSE,
  zoom.ratio = 2, zoom.suffix = " (zoom)", base.color = "white",
  highlight.color = "grey80", point.size = 0.5, color.sex.bias = TRUE,
  sex.bias.palette = c("firebrick1", "black", "dodgerblue2"),
  color.unmapped = TRUE, unmapped.palette = c(`0` = "dodgerblue3", `1` =
  "goldenrod1", `2` = "grey30"), signif.threshold = 0.05,
  sector.titles.expand = 1.9)
}
\arguments{
\item{mapping_file_path}{Path to a mapping results file generated by RADSex map.}

\item{contig_lengths_file_path}{Path to a contig lengths file generated by RADSex map.}

\item{chromosomes_names_file_path}{Path to a chromosomes names file, i.e. a tabulated file with name in the reference genome file as the first column and
corresponding chromosome name (starting with "LG" or "chr", case unsensitive) as the second column. If the chromosomes names in the reference genome file
already start with "LG" or "chr", there is no need for a chromosomes names file (default NULL).}

\item{plot.unplaced}{If TRUE, unplaced contigs will be plotted as a supercontig (default TRUE).}

\item{output_file_path}{Path to the plot output file. If the output file is not specified, the circular plot will be plotted in the default device (default NULL).}

\item{width}{Width of the output file in pixels (default 1400).}

\item{height}{Height of the output file in pixels (default 1400).}

\item{res}{Resolution of the output file in \% (default 100).}

\item{zoom.highlights}{If TRUE, highlighted scaffolds will be zoomed on at the top of the plot (default FALSE).}

\item{zoom.ratio}{Zooming factor for highlighted scaffolds, i.e. a size multiplier (default 2)}

\item{zoom.suffix}{A suffix to add after the name of a zoomed scaffold (default " (zoom)")}

\item{base.color}{Background color of a standard sector of the plot (default "white").}

\item{highlight.color}{Background color of a highlighted sector of the plot (default "grey80").}

\item{point.size}{Size of a point in the plot (default )}

\item{color.sex.bias}{If TRUE, points on the sex-bias track will be colored according to sex.bias.palette (default TRUE).}

\item{sex.bias.palette}{A vector of three colors defining the sex-bias track palette: female-biased, neutral, male-biased. (default c("firebrick1", "black", "dodgerblue2"))}

\item{color.unmapped}{If TRUE, unmapped scaffolds will be colored with alternating colors, similar to a manhattan plot (default TRUE).}

\item{signif.threshold}{Significance threshold for association with sex (default 0.05).}

\item{highlights}{A vector of scaffolds to highlight: c("LG1", "LG7", "scaffold_8") ... (default NULL).}

\item{unmapped.paltte}{A named vector of three colors: "0" = alternating color 1, "1" = alternating color2, and "2" = color for mapped scaffolds (default c("0"="dodgerblue3", "1"="goldenrod1", "2"="grey30")).}

\item{sector.title.expand}{A factor that controls the distance between sector titles and sector top axes (default 1.9).}
}
\description{
Generates a circular plot of RADSex mapping results in which each sector represents a linkage group and
the x-axis represents the position on the linkage group. The y-axis on the first track shows the sex-bias of a sequence, and the second track shows the
probability of association with sex of a sequence.
}
\examples{
plot_genome(mapping_file_path, contig_lengths_file_path, chromosomes_names_file_path = chromosomes_names_file_path, plot.unplaced=TRUE,
            output_file_path = NULL, width = 1400, height = 1400, res = 100,
            highlight=NULL, zoom.highlights = FALSE, zoom.ratio = 2, zoom.suffix = " (zoom)",
            base.color = "white", highlight.color = "grey80", point.size = 0.5,
            color.sex.bias = TRUE, sex.bias.palette = c("firebrick1", "black", "dodgerblue2"),
            color.unmapped = TRUE, unmapped.palette = c("0"="dodgerblue3", "1"="goldenrod1", "2"="grey30"),
            signif.threshold = 0.05, sector.titles.expand = 1.9)

}
