% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_manhattan.R
\name{plot_genome_manhattan}
\alias{plot_genome_manhattan}
\title{Mapping manhattan plot}
\usage{
plot_genome_manhattan(mapping_file_path, contig_lengths_file_path,
  chromosomes_names_file_path = NULL, plot.unplaced = TRUE,
  output_file_path = NULL, width = 14, height = 7, dpi = 300,
  point.size = 0.5, signif.threshold = 0.05,
  point.palette = c("dodgerblue3", "darkgoldenrod2"),
  background.palette = c("grey85", "grey100"),
  significance.line.color = "black", significance.line.type = 2,
  significance.text.position = c(0.05, 0.05))
}
\arguments{
\item{mapping_file_path}{Path to a mapping results file generated by RADSex map.}

\item{contig_lengths_file_path}{Path to a contig lengths file generated by RADSex map.}

\item{chromosomes_names_file_path}{Path to a chromosomes names file, i.e. a tabulated file with name in the reference genome file as the first column and
corresponding chromosome name as the second column. If the chromosomes names in the reference genome file start with "LG", "NC", or "CHR" (case unsensitive),
chromosomes can be detected automatically (default NULL).}

\item{plot.unplaced}{If TRUE, unplaced contigs will be plotted as a supercontig (default TRUE).}

\item{output_file_path}{Path to the plot output file. If the output file is not specified, the circular plot will be plotted in the default device (default NULL).}

\item{width}{Width of the output file in inches (default 14).}

\item{height}{Height of the output file in inches (default 7).}

\item{dpi}{Resolution of the output file (default 300).}

\item{point.size}{Size of a point in the plot (default 0.5)}

\item{signif.threshold}{Significance threshold for association with sex (default 0.05).}

\item{point.palette}{Color palette for the dots (default c("dodgerblue3", "darkgoldenrod2"))}

\item{background.palette}{Color palette for the background (default c("grey85", "grey100"))}

\item{significance.line.color}{Color for significance line, set to NULL for no line (default "black").}

\item{significance.line.type}{Linetype for the significance line, as usually defined in R (default 2).}

\item{significance.text.position}{X and Y axis offset for the significance text, as fractions of total axis length (default c(0.05, 0.05)).}
}
\description{
Generates a manhattan plot of association with sex from RADSex mapping results.
}
\examples{
plot_genome_manhattan("mapping_results.tsv", "contig_lengths.tsv", chromosomes_names_file_path = "chromosomes_names.tsv",
                      output_file_path = "mapping_results.png", point.size = 1,
                      signif.threshold = 0.01, point.palette = c("red3", "dodgerblue2"))

}
