% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sex_distribution.R
\name{plot_sex_distribution}
\alias{plot_sex_distribution}
\title{Plot sex distribution}
\usage{
plot_sex_distribution(input_file_path, output_file_path = NULL,
  title = NULL, width = 8, height = 6, dpi = 300, autoscale = FALSE,
  show.significance = TRUE, significance.color = "red3",
  significance.threshold = 0.05, color.scale.bins = c(0, 1, 5, 25, 100,
  1000), color.scale.colors = c("white", "navyblue"))
}
\arguments{
\item{input_file_path}{Path to a table of distribution of sequences between sexes.}

\item{output_file_path}{Path to the plot output file. If no output file is specified, the heatmap will be plotted in the default device.}

\item{title}{Plot title.}

\item{width}{Width of the output file (in inches).}

\item{height}{Height of the output file (in inches).}

\item{dpi}{Resolution of the output file.}

\item{autoscale}{If set to TRUE, the height of the output file will be automatically set so that horizontal and vertical scales are equal.
The resulting heatmap will not be square if the number of males and females in the population are very different.}

\item{show.significance}{If TRUE, tiles with significant association with sex are highlighted with the color defined in the significance.color parameter.}

\item{significance.color}{Color of the border for tiles significantly associated with sex.}

\item{significance.threshold}{P-value threshold to consider a tile significantly associated with sex.}

\item{color.scale.bins}{A vector of values to use as bins in the color palette.}

\item{color.scale.colors}{A vector of two colors used to create the color palette gradient.}
}
\value{
A heatmap stored in a ggplot object.
}
\description{
Plots the heatmap of distribution of sequences between sexes from a table generated with RADSex.
}
\examples{
# No autoscale: the plot dimensions are specified with "width" and "height".
heatmap = sex_distribution_heatmap(input_file_path, output_file_path='heatmap.png',
                                   title="Distribution of sequences between sexes",
                                   width=10, height=10, dpi=200,
                                   significance.color="gold", significance.threshold=0.05,
                                   color.scale.bins=c(0, 10, 100, 1000),
                                   color.scale.colors=c("white", "grey10"))

# Autoscale: the plot dimensions are automatically calculated from "width".
heatmap = sex_distribution_heatmap(input_file_path, output_file_path='heatmap.png',
                                   title="Distribution of sequences between sexes",
                                   width=10, autoscale=TRUE)
}
