% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{coverage_clustering}
\alias{coverage_clustering}
\title{Cluster coverage data}
\usage{
coverage_clustering(data, min.coverage = 0, max.coverage = 150,
  distance.method = "euclidean", clustering.method = "ward.D")
}
\arguments{
\item{data}{A table of coverage obtained with the \code{\link{load_coverage_table}} function.}

\item{min.coverage}{Minimum coverage value to consider a sequence present in an individual: coverage lower than this value will be set to 0 (default 0).}

\item{max.coverage}{Maximum coverage allowed in an individual: coverage higher than this value will be set to this value (default 100).}

\item{distance.method}{Method to use to compute the distance matrix.
Possible values: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". See \code{\link[stats]{dist}} for details (default euclidean).}

\item{clustering.method}{Method to use in the clustering.
Possible values: "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).
See \code{\link[stats]{hclust}} for details (default ward.D).}
}
\value{
A list with the following elements:
\item{data}{A data frame of coverage with individuals and sequences ordered based on the clustering results}
\item{individuals}{Individuals clustering results}
\item{sequences}{Sequences clustering results}
\item{distribution}{Distribution of coverage values}
}
\description{
Cluster individuals and sequences in a coverage table.
}
\examples{
clustering_data <- coverage_clustering(data,
                                       min.coverage = 0, max.coverage = 100,
                                       distance.method = "binary", clustering.method = "complete")
}
