% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_coverage_table.R
\name{load_coverage_table}
\alias{load_coverage_table}
\title{Load a coverage table}
\usage{
load_coverage_table(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a table of coverage.}
}
\value{
A data frame containing the coverage data.
}
\description{
Loads a table of coverage obtained with RADSex subset, RADSex signif, or RADSex loci.
}
\examples{
data <- load_coverage_table("coverage_table.tsv")
}
