% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_popmap.R
\name{load_popmap}
\alias{load_popmap}
\title{Load a population map}
\usage{
load_popmap(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a population map.}
}
\value{
A vector with individuals as names and sexes as values.
}
\description{
Loads a population map from a popmap file.
}
\examples{
popmap <- load_popmap("popmap.tsv")
}
