% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sex_distribution_table.R
\name{load_sex_distribution_table}
\alias{load_sex_distribution_table}
\title{Sex distribution table loader}
\usage{
load_sex_distribution_table(input_file_path)
}
\arguments{
\item{input_file_path}{Path to a table of distribution of sequences between sexes.}
}
\value{
A list object containing the data, number of males, and number of females.
}
\description{
Load a table of distribution of sequences between sexes obtained with RADSex sex_distribution.
}
\examples{
data = load_sex_distribution_table(input_file_path)
}
