% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage}
\usage{
plot_coverage(input_file_path, output_file_path = NULL,
  popmap_file_path = NULL, width = 10, height = 8, dpi = 300,
  title = NULL, min.coverage = 0, max.coverage = 150,
  distance.method = "euclidean", clustering.method = "ward.D",
  males.color = "dodgerblue3", females.color = "red3",
  coverage.palette = c("white", "royalblue2", "black", "gold2", "red3"),
  individual.names = TRUE, sequence.names = FALSE,
  individual.dendrogram = TRUE, sequence.dendrogram = TRUE)
}
\arguments{
\item{input_file_path}{Path to a coverage table generated by RADSex.}

\item{popmap_file_path}{Path to a population map file.}

\item{width}{Width of the output file (in inches).}

\item{height}{Height of the output file (in inches).}

\item{dpi}{Resolution of the output file.}

\item{title}{Plot title.}

\item{min.coverage}{Minimum coverage value to consider a sequence present in an individual. Coverage lower than min.coverage will be set to 0.}

\item{max.coverage}{Maximum sequence coverage allowed in an individual. Coverage higher than max.coverage will be set to max.coverage.}

\item{distance.method}{Method to use to compute the distance matrix.
Possible values: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". See \code{\link[stats]{dist}} for details.}

\item{clustering.method}{Method to use in the clustering step.
Possible values: "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).
See \code{\link[stats]{hclust}} for details.}

\item{males.color}{If a popmap is specified, sets the color of male individual names on the horizontal axis.}

\item{females.color}{If a popmap is specified, sets the color of female individual names on the horizontal axis.}

\item{coverage.palette}{Color palette for coverage. The value should be a vector of length 5 corresponding to the following intervals/values:
(0, 1 : mean, mean : 3rd quartile, 3rd quartile : (max - 1), max)}

\item{individual.names}{If TRUE, shows individual names on the x-axis.}

\item{sequence.names}{If TRUE, shows sequence names on the y-axis.}

\item{individual.dendrogram}{If TRUE, shows individual clustering dendrogram on the x-axis.}

\item{sequence.dendrogram}{If TRUE, shows sequence clustering dendrogram on the y-axis.}
}
\description{
Plots a heatmap of coverage from a table generated with RADSex subset, RADSex signif, or RADSex loci.
}
\examples{
plot_coverage(input_file_path, popmap_file_path="./popmap.tsv",
              width=12, height=10,
              title="Individuals and sequences clustering based on coverage",
              min.coverage=5,
              clustering.method="ward.D2", distance.method="binary",
              males.color="blue", females.color="red",
              coverage.palette=c("white", "green", "black", "red", "red3"),
              individual.names=TRUE, sequence.names=TRUE,
              individual.dendrogram=TRUE, sequence.dendrogram=TRUE)
}
